shiny_app/requirements.txt
# FDA Food Adverse Event Shiny App – dependencies
# Install: pip install -r requirements.txt

shiny>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
